changed way objects are constructed
Fixed warning in firexfox... mutating the [[Prototype]] of an object will cause your code to run very slowly; instead create the object with the correct initial [[Prototype]] value using Object.create...

I create a constructor function every time an entity is being Entitize()'d. I then set the new constructor's prototype property to the base entity object for the provided type. An empty object is created with this constructor, and the JEntity constructor is called using call/apply, setting the context to the newly created empty object. This allowed me to avoid explicitly setting the __proto__ property in the JEntity constructor when the new instance was being created.